add sol to 122 Hi admin:
I want to pay attention to the rational regarding to problem of  best time to buy. Normally, it is the ratio of selling ratio price to purchasing price instead of difference between selling price and purchasing price. For example, buy at 10 and sell at 100 is equal to buying at 1 and selling at 10. Right?

log each element in this prices array before do the math can fix it.

Thanks